Change mmsi parsing from i32 to u32 and database type to BIGINT
The mmsi parsing has been updated in csvreader.rs from i32 to u32 to prevent accepting negative numbers. Simultaneously, the database field type for mmsi and imo has been changed to BIGINT in both dynamic and static tables to support larger ranges.

+++ aisdb/aisdb_sql/psql_createtable_static.sql
@@ -1,10 +1,10 @@
 CREATE TABLE IF NOT EXISTS ais_{0}_static (
-    mmsi INTEGER NOT NULL,
+    mmsi BIGINT NOT NULL,
     time INTEGER NOT NULL,
     vessel_name TEXT,
     ship_type INTEGER,
     call_sign TEXT,
-    imo INTEGER NOT NULL DEFAULT 0,
+    imo BIGINT NOT NULL DEFAULT 0,
     dim_bow INTEGER,
     dim_stern INTEGER,
     dim_port INTEGER,
